Move game and bot list to their classes
