+++ src/bot_discord.ts
